+++ src/contracts/scaffs-config.ts
@@ -8,16 +8,18 @@
  */
 export interface ScaffsConfig {
     scaffsPaths: string[];
-    scaffs: {[key: string]: string};
+    scaffs: { [key: string]: string };
 
     // used to load scaffolds by name (can be populated with)
-    absoluteScaffPaths: {[key: string]: string};
+    absoluteScaffPaths: { [key: string]: string };
 
     // used to resolve absolute paths
     baseConfigPath: string;
 
     // header text to prepend to all generated files
-    header?: string;
+    headers?: {
+        [glob: string]: string;
+    };
 
     // change the scaffs default intention to the provided characters
     indention?: string;