+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.5.0 - BREAKING CHANGE
+
+* Remove `header` (singular) from scaffs-config
+* Add `headers` to scaffs-config to allow different headers based on glob matchers
+* Minor version bump because no one is using this package yet
+
 ## 0.4.0
 
 * Automatic discovery of registry scaffs (@scaffs)